Add transformer to calculate rolling features
Summary:
Creates an sklearn transformer to do the task that was achieved in R by the 'buildFeaturesRolling' function
this transformer operates on a pandas dataframe, and applies a rolling function on it.

For now, there is no 'wrapper' yet, but this can be written later. It basically involves spreading the columns, and temporarily removing the time/type/id columns

Test Plan: tested on iris dataset, no unit tests yet

Reviewers: Sebastiaan

Reviewed By: Sebastiaan

Subscribers: rutger, tim

Differential Revision: https://dev.ynformed.nl/D366