Added docs/build to .gitignore
Summary:
Documentation is making sam diffs hard to parse. also, documentation in the repo is hard to access for outsiders.
a solution that solves both problem is having jenkins build the documentation and host it externally. This has been done: T331
Now that this is done, the documentation can be removed from the repo. Also, it can be added to .gitignore. This means people can still build
documenation locally if they want, it just won't affect the repo. docs/source will still need to be updated as always.

Test Plan: It's a .gitignore change, no tests needed really

Reviewers: tim, Sebastiaan, rutger

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D445