Update documentation, remove some useless functions, and fix the many, many warnings
Summary:
Documentation didn't work for visualization. This turned out to be because the 'tkinter' package was required. But this was obscured because of the hundreds of warnings.
This commit gets rid of almost all the warnings. I installed tkinter on ynfsrv01, but that is not part of this commit. However, these kinds of issues should be far easier to track down when you can actually read the warnings

Test Plan: Ran documentation locally, it worked.

Reviewers: tim, Sebastiaan

Reviewed By: tim

Differential Revision: https://dev.ynformed.nl/D456