Undo a mistake from last commit, some warnings return
Summary:
last commit removed a lot of warnings, but unfortunately also removed the member table for classes. Since warnings aren't shown to users anyway, this commit returns the member table
Hopefully in the future there can be a way to remove warnings as well as still have the table. But after at least half an hour of googling, i haven't found it yet

Test Plan: Tested it locally, it works

Reviewers: tim, Sebastiaan

Reviewed By: tim

Differential Revision: https://dev.ynformed.nl/D458