Fix bug when gap happens on beginning/end of input
Summary:
There was a bug when there was a gap on the beginning/end.
This was because a 'gap' is defined as consecutive non-outliers, however it's not a gap if it occurs at the beginning/end.

This bug is now fixed, and new unit tests were written for gap size 1 and 2.

Test Plan: New unit tests succeed. Also ran pytest, many less warnings (because pytest previously threw settingwithcopywarnings)

Reviewers: tim, rutger

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D485