Add unit test for MongoWrapper
Summary:
Added unit tests. also made them automatically skip if an initial connection cannot be made within 100 milliseconds. It should be visible that the tests were skipped. But of course, if there were no changes to the mongo code, it doesn't matter if it was skipped or not. This makes development easier on a machine with no mongodb connection.
This commit also shows how to make a custom pytest decorator. Decorators ftw!

Test Plan: Ran unit tests locally, and it worked. I also (hackily) changed localhost in the code to 'foobar'. The unit tests were indeed skipped as expected. I did not test on a device with no mongo, but I would expect it to work the same.

Reviewers: rutger, tim

Reviewed By: tim

Differential Revision: https://dev.ynformed.nl/D496