Add unit test for rolling features
Summary:
Added unit tests for rolling features. Also added clarification for the function where I thought needed.
Not every single combinations of inputs are tested because that's way too much. But I'm pretty sure the important stuff is tested. As explained, std, var, etc aren't tested because their code path is exactly the same as sum, mean, etc.
Functionally nothing changed, except for a strange typo bug that caused fourier not to work entirely. That was an easy fix. Also a simple assertion for negative lookback because that is undefined, and don't want to give the impression that negative lookbacks are supported. Maybe we can support that later if there is demand.

fixes T328

Test Plan: Ran unit tests

Reviewers: rutger, tim

Reviewed By: rutger

Maniphest Tasks: T328

Differential Revision: https://dev.ynformed.nl/D499