Update mongo unittest to be more robust
Summary:
mongo unit test fails on different hosts than ynfsrv01. The unit test should work on different development environments (as long as those development environments have mongo in the first place of course, otherwise the mongo_wrapper is useless to begin with.
This update changed it so that it should work on more environments than just ynfsrv01.

Test Plan: ran it on jupyter-sam.ynformed.nl and ynfsrv01, it worked on both.

Reviewers: rutger

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D537