Add sum_grouped_columns function
Summary:
Add a new function that sums columns based on their group. This is a utility function that will be needed when T292 finally gets solved (which will be done through documentation)

It ended up being a bit overkill, since the funciton is only a few lines... but now that it's done, oh well.

Test Plan: Unit tests, written, they work

Reviewers: rutger, tim

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D558