Fix many flake8 (and pep8) violations
Summary: Also add default flake8 configuration (not invoked by arc lint, but may be useful in the future if someone wants to flake8 their code)

Test Plan:
Unit tests still succeed, many less flake8 violations (from 100+ to only 20 now)

The remaining 20 are kinda stupid:

imported but unused in __init__.py... which is by design so didn't fix those.

also using \d and \w in documentation. (they are not allowed in code but it's not smart enough to see that it's in documentation

Reviewers: rutger

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D560