Overhaul readme, and fix several documentation issues
Summary:
Fix T408, T396, T394, T392

The readme has not been updated in a while, so I decided there could be a more logical ordering. Now, at the top there is 'getting started', showing how to install it and how to see documentation. Below that is all the stuff for developers.

Also fix a few other bugs in the documentation

Lastly, I fixed the documentation saying to use pep8. See https://github.com/PyCQA/pycodestyle/issues/466 It is now named pycodestyle (pep8 refers to the guide, not the program used to enforce the guide) pep8 is kept around, because arc lint looks at the [pep8] section, not the [pycodestyle] option

Test Plan: Haven't tested it, but they are simple changes to it should work

Reviewers: rutger, tim

Reviewed By: rutger

Maniphest Tasks: T408, T396, T394, T392

Differential Revision: https://dev.ynformed.nl/D581