Remove redundant options from rolling features, add checks for timeoffset.
Summary:
For now, timeoffset is only supported if the dataframe has a timeindex. Will probably try adding an extra time column in the future.
But this is a good first step. (and I think it's better to keep diffs small)

I also removed a bunch of checks on the type of window_size. This was needed because it was too strict, and it wasn't correct anyway. It only accepted (int, float), but for example not np.int32. So it was kind of pointless to keep adding more numeric datatypes (it was also missing long), especially when python is ducktyped anyway...

Test Plan: All unit tests succeed

Reviewers: tim, rutger, daanvanes

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D703