Hide optional dependencies, fixing T580
Summary:
Previously, when not having knmy or pymongo installed, there was the possibility of the package randomly crashing when it required a package that was not installed, evne if the package was unneeded.
The worst example was pymongo, which was required by utils, which was required by a lot of other stuff. Even though pymongo was only required for a single, rarely-used function.

Now, everything is hidden so even if you don't have pymongo installed, every other function wil lstill work.
The only stuff that isn't hidden are standard library (logging, math, etc), and the three main dependencies: pandas, numpy, scikit-learn

Test Plan: Unit tests still work, and they load the required packages lazily (with this I mean: the unit tests don't know to load pymongo until they see 'import pymongo' only when neccecary)

Reviewers: rutger

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D739