Update documentation to 1.0
Summary:
see title

this is just the FIX of the old deprecated functions and a bunch of syntax errors that caused documentation to look weird.

Not yet a thorough runthrough to see if all documentation is actually corret. That is (T 590). This is T609.

Also, miinimum viable project is not updated, that is a lot of work, and I want to keep the diffs small.

Test Plan: It runs on my laptop, come visit me to check :)

Reviewers: rutger, Loes, Sebastiaan

Reviewed By: Sebastiaan

Differential Revision: https://dev.ynformed.nl/D776