Overhaul correct_extremes, fixing T624, T621
Summary:
Kinda changes the function quite a bit, to accept series instead of dataframe. also missing values are not seen as outliers.
Maybe in the future we can add option 'missing_values_are_outliers' which would be simple: just set na_locations to [] if that option is true
But that would be an enhancement, so not neccecary for now

commit to fix arc lint warnings

Test Plan: Unit tests succeed

Reviewers: rutger, tim

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D859