Fix importing tensorflow and using gpu for unit tests
Summary:
Tensorflow was imported even though it is an optional dependency.
Fixed it.
Unit tests didn't run if GPU was not available even though the unit tests are so minimal they can run perfectly on CPU
Disable gpu for unit tests

Test Plan: It's simple enough that I didn't test it. unit tests run even when Med is using 100% of the GPU memory

Reviewers: arjan, daanvanes, rutger

Reviewed By: arjan

Differential Revision: https://dev.ynformed.nl/D1201