Fix bug in incident heatmap, T815
Summary:
The bug was that seaborn set the axes to (2.5, 0.5) even though it should be (0, 3) or (3, 0)
I put a round/ceil in there
I also changed the unit tests to not be so ridiculously ugly. They actually look like normal plots now.

Test Plan:
it works on the matrix.

EDIT: Tested it on my laptop (windows), unit tests succeeded there as well

Reviewers: daanvanes, rutger

Reviewed By: rutger

Differential Revision: https://dev.ynformed.nl/D1204