Add benchmarking functionality
Summary:
D1242
Add some functions for benchmarking. Not sure if they are really usable by random people, for me they seem to work fine, for now.

Test Plan: No unit tests, but I'm not really sure if that is worth it for these... They are kinda wrapper functions around a bunch of sam stuff, I'm not sure if writing a unit test really adds anything, especially if they won't be used by other people anyway. It's more of a 'run' kind of thing.

Reviewers: daanvanes, Loes

Reviewed By: daanvanes

Subscribers: arjan

Differential Revision: https://dev.ynformed.nl/D1257