Add standard scaler and return history
Summary:
the model does much, much, much better when we enable scaling. It goes from 'always predict 0 which is the same as persitence benchmark' to actually performing significantly better than the persistence benchmark.

This also is 'neccecary' for the demo, because it looks unimpressive to have a model that does the same as the persistence benchmark. This shows it can actually work in principles

Test Plan: it works locally for me, MSE went from 0.78 (same as persistence) to 0.37

Reviewers: daanvanes, Loes

Reviewed By: daanvanes

Differential Revision: https://dev.ynformed.nl/D1262